+++ CHANGES.rst
@@ -14,7 +14,11 @@
   reflect its purpose.
 - Make ``PersistentFloatTo64BitIntNormalizer`` subclass
   ``FloatTo64BitIntNormalizer``.
-
+- Add ``IDeferredCatalog`` and an implementation in
+  ``DeferredCatalog`` to allow creating catalog objects that don't
+  participate in event subscription-based indexing. This replaces
+  ``IMetadataIndex``, which is now an alias for this object. See
+  `issue 3 <https://github.com/NextThought/nti.zope_catalog/pull/3>`_.
 
 1.0.0 (2017-06-15)
 ==================